fix: stop forcing shipping_fee to float, breaking real DynamoDB writes
create_delivery_order() cast shipping_fee to float(), which made
shipping_fee_amount/total_amount native Python floats. Real DynamoDB
(via boto3) rejects float values outright ("Float types are not
supported. Use Decimal types instead."), so every order submission —
both the wizard and the chat flow — failed at the STORE.save_request()
call whenever running against the real backend (USE_MOCK=false). The
local MemoryStore used by the test suite doesn't enforce this
constraint, so none of the existing tests caught it.

All prices in the static store catalog are whole numbers, so this
just drops the float() cast — the whole amount calculation now stays
int end to end, sidestepping the issue without introducing
decimal.Decimal handling.

+++ backend/app/services/delivery.py
@@ -111,7 +111,7 @@ def _validate_payload(payload: dict) -> dict | None:
     return None
 
 
-def calculate_order_amounts(goods: list, shipping_fee: float = 0) -> dict:
+def calculate_order_amounts(goods: list, shipping_fee: int = 0) -> dict:
     """Calculate order totals."""
     original_amount = sum(item.get("price", 0) * item.get("quantity", 1) for item in goods)
     return {
@@ -135,8 +135,9 @@ def create_delivery_order(actor_id: str, payload: dict) -> dict:
     note = payload.get("note", "")
     store_url = payload.get("store_url", "")
 
-    # 計算金額
-    shipping_fee = float(payload.get("shipping_fee", 60))
+    # 計算金額（一律用 int，DynamoDB 不支援 Python float，見
+    # test_delivery_service.py 的迴歸測試）
+    shipping_fee = int(payload.get("shipping_fee", 60))
     amounts = calculate_order_amounts(goods, shipping_fee)
 
     order_items = {
